main: Implement argument parsing
Implement argument parsing, using arguments supplied both in the command
line as well as arguments supplied in a file.

Considering that most users will simply want to launch the application and
have it working, the application tries to read a configure file regardless
of whether or not it was specified. Otherwise, users would need to create a
script to launch the application specifying the configuration file.