main: Implement clearing a game
Implement DELETE in a game, without specifying a resource, to remove every
resource associated with the given game. Since there isn't any resource
associated with this object, a new field 'cmd' was added to the object sent
by the WebSocket to specify whether a resource was modified (by setting
'cmd' to 'SET') or whether the entire game was cleared (by setting 'cmd' to
'CLEAR').

On the Javascript side, the function ws_tracker.setClearFunction() must be
called to specify which function should be called in response to a 'CLEAR'
'cmd' sent over WebSocket.